requirements.txt
requests
beautifulsoup4
lxml
duckduckgo-search
wikipedia
rich
streamlit
tqdm
psutil
python-dotenv
plotly
scipy
pandas
plotly
numpy
scipy
pytz
anthropic
feedparser
joblib
lightgbm
logzero
openai
pyotp
scikit-learn
tensorflow
textblob
trafilatura
websocket-client
websockets
xgboost
yfinance
PyYAML
ta
smartapi-python
psutil>=5.9.0
plotly
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
numpy>=1.26.0
